packages/glife/glife-0.2.tar.gz/glife-0.2/src/glife/gol/tests/test_api.py

import pytest
from ..api import GameOfLife
from ..rules.conway import ConwayRules

@pytest.fixture
def game_of_life_api():
    return GameOfLife(5, 5, ConwayRules())


def test_api_01(game_of_life_api: GameOfLife):
    game_of_life_api._field.init([
        [0, 0, 0, 0, 0],
        [0, 1, 1, 1, 0],
        [0, 0, 1, 0, 0],
        [0, 0, 0, 0, 0],
        [0, 0, 0, 0, 0],
    ])
    expected = [
        [0, 0, 1, 0, 0],
        [0, 1, 1, 1, 0],
        [0, 1, 1, 1, 0],
        [0, 0, 0, 0, 0],
        [0, 0, 0, 0, 0],
    ]

    ng = game_of_life_api.next_generation()
    assert expected == ng._field

def test_api_02(game_of_life_api: GameOfLife):
    game_of_life_api._field.init([
        [0, 0, 1, 0, 0],
        [0, 1, 1, 1, 0],
        [0, 1, 1, 1, 0],
        [0, 0, 0, 0, 0],
        [0, 0, 0, 0, 0],
    ])
    expected = [
        [0, 1, 1, 1, 0],
        [0, 0, 0, 0, 0],
        [0, 1, 0, 1, 0],
        [0, 0, 1, 0, 0],
        [0, 0, 0, 0, 0],
    ]

    ng = game_of_life_api.next_generation()
    assert expected == ng._field


def test_api_03(game_of_life_api: GameOfLife):
    game_of_life_api._field.init([
        [0, 1, 1, 1, 0],
        [0, 0, 0, 0, 0],
        [0, 1, 0, 1, 0],
        [0, 0, 1, 0, 0],
        [0, 0, 0, 0, 0],
    ])
    expected = [
        [0, 0, 1, 0, 0],
        [0, 1, 0, 1, 0],
        [0, 0, 1, 0, 0],
        [0, 0, 1, 0, 0],
        [0, 0, 0, 0, 0],
    ]

    ng = game_of_life_api.next_generation()
    assert expected == ng._field


def test_api_04(game_of_life_api: GameOfLife):
    game_of_life_api._field.init([
        [0, 0, 1, 0, 0],
        [0, 1, 0, 1, 0],
        [0, 0, 1, 0, 0],
        [0, 0, 1, 0, 0],
        [0, 0, 0, 0, 0],
    ])
    expected = [
        [0, 0, 1, 0, 0],
        [0, 1, 0, 1, 0],
        [0, 1, 1, 1, 0],
        [0, 0, 0, 0, 0],
        [0, 0, 0, 0, 0],
    ]

    ng = game_of_life_api.next_generation()
    assert expected == ng._field


def test_api_05(game_of_life_api: GameOfLife):
    game_of_life_api._field.init([
        [0, 0, 1, 0, 0],
        [0, 1, 0, 1, 0],
        [0, 1, 1, 1, 0],
        [0, 0, 0, 0, 0],
        [0, 0, 0, 0, 0],
    ])
    expected = [
        [0, 0, 1, 0, 0],
        [0, 1, 0, 1, 0],
        [0, 1, 0, 1, 0],
        [0, 0, 1, 0, 0],
        [0, 0, 0, 0, 0],
    ]

    ng = game_of_life_api.next_generation()
    assert expected == ng._field
